en masse rettelser blandt andet til login-proceduren og til popop-javescripts

+++ src/main/java/app/controllers/ItemController.java
@@ -83,20 +83,10 @@ public class ItemController {
     }
 
     private static void createOrder(Context ctx, ConnectionPool connectionPool) throws DatabaseException {
-        User currentUser = ctx.sessionAttribute("currentUser");
-        if (currentUser == null) {
-            ctx.render("login.html");
-            return;
-        }
 
         if (ctx.sessionAttribute("orders") != null) {
             orderLine = ctx.sessionAttribute("orders");
         }
-
-        String email = currentUser.getEmail();
-        String name = currentUser.getName();
-        String mobile = currentUser.getMobile();
-        int balance = currentUser.getBalance();
 
         int toppingId = Integer.parseInt(ctx.formParam("topping")); // Assumes you have toppingId in form
         int bottomId = Integer.parseInt(ctx.formParam("bund"));   // Assumes you have bottomId in form
@@ -110,23 +100,20 @@ public class ItemController {
 
         int orderlinePrice = calculateOrderLinePrice(topping, bottom, quantity);
 
-        Order order = new Order(currentUser.getUserId(), email, name, mobile, balance, topping.getTopping(), bottom.getBottom(), quantity, orderlinePrice);
+        Order order = new Order(topping.getTopping(), bottom.getBottom(), quantity, orderlinePrice);
         orderLine.add(order);
 
         ctx.sessionAttribute("orders", orderLine);
 
-
         int totalAmount = 0;
         int orderCount = 0;
         for (Order orderline : orderLine) {
-            if (order.getUserId() == currentUser.getUserId()) {
-                totalAmount += orderline.getOrderlinePrice();
-                orderCount++;
-            }
+            totalAmount += orderline.getOrderlinePrice();
+            orderCount++;
         }
 
         ctx.sessionAttribute("totalAmount", totalAmount); // Sender det samlede beløb som en attribut til HTML-skabelonen
-        ctx.sessionAttribute("orderCount", orderCount);
+        ctx.sessionAttribute("orderCount", orderCount); // Sender det samlede antal ordrelinier til HTML-skabelonen
 
         showTopping(ctx, ConnectionPool.getInstance());
         showBottom(ctx, ConnectionPool.getInstance());
@@ -151,45 +138,58 @@ public class ItemController {
 
     public static void payForOrder(Context ctx, ConnectionPool connectionPool) throws DatabaseException {
         User currentUser = ctx.sessionAttribute("currentUser");
+        if (currentUser == null) {
+            ctx.sessionAttribute("fromcheckout", true);
+            ctx.render("login.html");
+            return;
+        }
+
         int orderprice = ctx.sessionAttribute("totalAmount");
 
-        if (currentUser.getBalance() >= orderprice) {
-            ArrayList<Order> tempOrderLine = ctx.sessionAttribute("orders");
+        if (orderprice != 0) {
+            if (currentUser.getBalance() >= orderprice) {
+                ArrayList<Order> tempOrderLine = ctx.sessionAttribute("orders");
 
-            int generatedOrderId = ItemMapper.insertOrder(tempOrderLine.get(0).getUserId(), connectionPool);
-            List<Topping> toppingList = ItemMapper.showToppings(connectionPool);
-            List<Bottom> bottomList = ItemMapper.showBottoms(connectionPool);
+                int generatedOrderId = ItemMapper.insertOrder(tempOrderLine.get(0).getUserId(), connectionPool);
+                List<Topping> toppingList = ItemMapper.showToppings(connectionPool);
+                List<Bottom> bottomList = ItemMapper.showBottoms(connectionPool);
 
-            for (Order order : tempOrderLine) {
-                int toppingId = 0;
-                for (Topping topping : toppingList) {
-                    if (order.getTopping().equals(topping.getTopping())) toppingId = topping.getToppingId();
+                for (Order order : tempOrderLine) {
+                    int toppingId = 0;
+                    for (Topping topping : toppingList) {
+                        if (order.getTopping().equals(topping.getTopping())) toppingId = topping.getToppingId();
+                    }
+                    int bottomId = 0;
+                    for (Bottom bottom : bottomList) {
+                        if (order.getBottom().equals(bottom.getBottom())) bottomId = bottom.getBottomId();
+                    }
+
+                    ItemMapper.payForOrder(generatedOrderId, toppingId, bottomId, order.getQuantity(), order.getOrderlinePrice(), connectionPool);
+
                 }
-                int bottomId = 0;
-                for (Bottom bottom : bottomList) {
-                    if (order.getBottom().equals(bottom.getBottom())) bottomId = bottom.getBottomId();
-                }
-
-                ItemMapper.payForOrder(generatedOrderId, toppingId, bottomId, order.getQuantity(), order.getOrderlinePrice(), connectionPool);
-
+                tempOrderLine.clear();
+                ItemMapper.deleteUsersBasket(currentUser.getUserId(), connectionPool);
+                int newBalance = currentUser.getBalance() - orderprice;
+                currentUser.setBalance(newBalance);
+                UserMapper.updateBalance(currentUser.getUserId(), newBalance, connectionPool);
+                ctx.attribute("message", "Tak for din ordre. Din ordre har fået ordrenummer " + generatedOrderId + ". Du hører fra os når din ordre er parat til afhentning!");
+                ctx.attribute("ordercreated", true);
+                ctx.sessionAttribute("totalAmount", 0);
+                ctx.sessionAttribute("orderCount", 0);
+                showTopping(ctx, ConnectionPool.getInstance());
+                showBottom(ctx, ConnectionPool.getInstance());
+                ctx.render("index.html");
+            } else {
+                ctx.attribute("message", "Din saldo lyder på " + currentUser.getBalance() + " kr. hvilket ikke er nok til at betale for ordren! Fjern nogle varer fra din kurv eller indbetal penge på din konto!");
+                ctx.attribute("notenoughtmoney", true);
+                ctx.render("checkoutpage.html");
             }
-            tempOrderLine.clear();
-            ItemMapper.deleteUsersBasket(currentUser.getUserId(), connectionPool);
-            int newBalance = currentUser.getBalance() - orderprice;
-            currentUser.setBalance(newBalance);
-            UserMapper.updateBalance(currentUser.getUserId(), newBalance, connectionPool);
-            ctx.attribute("message", "Tak for din ordre. Din ordre har fået ordrenummer " + generatedOrderId + ". Du hører fra os når din ordre er parat til afhentning!");
-            ctx.attribute("ordercreated", true);
-            ctx.sessionAttribute("totalAmount", 0);
-            ctx.sessionAttribute("orderCount", 0);
-            showTopping(ctx, ConnectionPool.getInstance());
-            showBottom(ctx, ConnectionPool.getInstance());
-            ctx.render("index.html");
         } else {
-            ctx.attribute("message", "Din saldo lyder på " + currentUser.getBalance() + " kr. hvilket ikke er nok til at betale for ordren! Fjern nogle varer fra din kurv eller indbetal penge på din konto!");
+            ctx.attribute("message", "Der er ingen varer i din kurv at betale for!");
             ctx.attribute("notenoughtmoney", true);
             ctx.render("checkoutpage.html");
         }
+
     }
 
 }
